pyproject.toml
[project]
name = "power-pole"
version = "0.9.1"
description = "Temperature time series forecasting with skforecast and sklearn (scikit-learn)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.2.1",
    "fastapi>=0.115.11",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "skforecast>=0.13.0",
    "toml>=0.10.2",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.3.5",
]
